Reviewed comments and added uint_16t type to port param
Done. I changed addListeningPort to use uint16_t as well. That keeps
the types consistent, though I believe it still would not prevent
silent wrapping at call sites when callers pass a wider integer type.

+++ ext/include/opentelemetry/ext/http/server/http_server.h
@@ -199,7 +199,7 @@ public:
 
   void setServerName(std::string const &name) { m_serverHost = name; }
 
-  int addListeningPort(int port)
+  int addListeningPort(uint16_t port)
   {
